moved search into subs. search input select all
