+++ src/Pages/Inventory/Components/PartItems/Search/Search.module.css
