+++ src/Pages/Inventory/Components/ToolItems/Search/Search.module.css
@@ -1,0 +1,15 @@
+.parent {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  flex-wrap: nowrap;
+  margin-top: 2%;
+  margin-right: 5%;
+}
+
+.container {
+  display: flex;
+  align-self: auto;
+  margin-right: 2%;
+  margin-bottom: 2%;
+}
